Added RestController for Person. Added person creating and deleting functions.

+++ src/main/java/ua/gradebook/model/dao/ContainerDAOImpl.java
@@ -30,7 +30,7 @@ public class ContainerDAOImpl implements DAO {
     }
 
     @Override
-    public ParentBean findById(Long id) {
+    public ParentBean findById(Integer id) {
         return null;
     }
