+++ src/main/java/ua/gradebook/model/dao/DisciplineDAOImpl.java
@@ -31,7 +31,7 @@ public class DisciplineDAOImpl implements DAO {
     }
 
     @Override
-    public ParentBean findById(Long id) {
+    public ParentBean findById(Integer id) {
         return null;
     }
